[XPTI][INFRA] Refactored API implementations to improve performance (#19160)
As a part of the performance study to improve SYCL runtime performance,
it was documented that XPTI when enabled by a tool or collector was
having higher than usual performance impact. This is the second PR to
address the specific xptiMakeEvent* API and current implementation is
~30% faster than previous implementation. The last PR in this series
will change the instrumentation in SYCL to address the large performance
impact caused by mixing debug and performance streams in the runtime
today.

 - Also fixes issues in XPTI samples due to changes in PR#18786

---------

Signed-off-by: Vasanth Tovinkere <vasanth.tovinkere@intel.com>