[SYCL][NFC] Split unit test kernels based on their args
Most of the unit test kernels don't have any arguments, and those that
do typically use a single accessor. Prior to this change, kernel size
was passed as a template parameter of the kernel function type,
while still sharing the same kernel name. This causes problems for the
upcoming change to kernel name based caches, which will be tied
to both types and strings.

Since currently there are only 2 cases for such test kernels: no arguments
and a single accessor, this patch simply adds a separate kernel for the
latter case.