[NFC][SYCL][Graph] Prepare for getSyclObjImpl to return raw ref (#19224)
I'm planning to change `getSyclObjImpl` to return a raw reference in a
later patch, uploading a bunch of PRs in preparation to that to make the
subsequent review easier.

I also did `s/sycl::detail::getSyclObjImpl/getSyclObjImpl/g` here to
increase readability and to make it simpler to perform the future patch
by simple `sed` (and avoid multiline changes).

This PR originated in `unittests/Extensions/CommandGraph`, but I had to
modify `isSimilar` and `hasSimilarStructure` in the implementation
itself too.