[SYCL][CMAKE] Switch back to _FORTIFY_SOURCE=2
The problem here is that the macro is actually handled by glibc and
value `3` isn't supported with older compiler/glibc combinations,
causing warnings about the macro redefinition.

We still have to support older compilers/glibc and therefore as a
temporary solution, reverting `_FORTIFY_SOURCE` back to `2` to avoid
build issues with `-Werror`.