[SYCL][Driver] Respect -nocudalib again after bad merge conflict resolution (#19216)
[fc26ef4cf0cafce381395f2d422bff33227e80b8][1] is a bad resolution
`--offloadlib` is a positive option to explicitly link offload device
libraries (usually on by default), while the option it replaced
(`-nogpulib`) was the negative option to disable linking them.
To behave correctly, we need to consider both positive and negative
options.

The test did not catch this because the filecheck CHECK-NOT directive
had a typo. Fix that as well.

[1]:
https://github.com/intel/llvm/commit/fc26ef4cf0cafce381395f2d422bff33227e80b8#diff-f8c64e36dfe3828a6f816c4550e78bb0305769ace1be53207e86ac9a3280ac9eR6291