[SYCL] Add standalone_debug attribute to sycl::device class (#19059)
A debug info optimization, known as "constructor homing," is enabled by
default and omits debug information for certain constructors based on
their characteristics. For the sycl::device class, the constructor is
neither a copy/move constructor nor deleted, and is non-trivial,
resulting in the omission of its debug information.
This omission causes the pretty printer to fail, as it requires the
debug information to display essential details such as types and device
names. To resolve this, the pull request introduces the standalone_debug
attribute to the sycl::device class, ensuring that debug information is
always emitted for this class.
While this change successfully addresses the pretty printer issue, it
may lead to debug information being present in every compilation unit,
potentially affecting code size. Testing with existing SYCL tests shows
that the code size impact is relatively small, with increases ranging
from approximately 0.6% to 1.5% in most cases.