Start to reduce dependency from libdevice to libclc
The libdevice project included a CMake file from libclc to complete some
work. This is an undesirable link for several reasons.

For one, libclc is upstream and libdevice isn't. This means that it's
easy to accidentally pull in changes from upstream that alter how
libdevice is built. We can see this in the FOLDER property of the
link_bc target, which was set to 'libclc/Device IR/Linking'. This would
be confusing for any user building DPC++ in an IDE.

For two, this sharing of code also necessitated making downstream
changes to libclc, which make maintenance more difficult. This can be
seen in the RSP_DIR argument to link_bc, which was never set in libclc
to anything other than the upstream value.

This commit therefore starts to sever this dependency by just copying
over the two utility functions to libdevice's CMake. They aren't very
big and it should make any existing dependencies clearer. For example,
we implicitly are relying on libclc being built so that the
prepare-bultins target and executable are set up. This isn't an ideal
situation but is left for future work.