[CI][Bench] Implement exponentially weighted moving average for SYCL nightly regression CI (#18766)
Median does not respond very fast to changes in performance, and thus is
not a suitable metric to be used for regression checking. This PR
implements an option to use exponentially weighted moving average
instead.

The hope is that this implementation could also be used on CPU
instructions retired metric when SYCL compute-benchmark tests start
reporting instructions retired; this would create a far more robust
metric to use in order to spot regressions.

**Note** to llvm-reviewers-benchmarking: Observe that changes here are
removed from the core benchmarking scripts -- This change should be
functionally NFC for the core benchmarking scripts.