[SYCL][CMAKE] Drop `nodlopen` from hardening flags (#19357)
We can't have this flag being applied globally, because it then affects
SYCL RT which has plenty of libraries that are supposed to be `dlopen`ed
by design.

It probably makes sense to apply `nodlopen` for some of our executables,
but that should be done on a case-by-case basis.

For now, let's just fix the toolchain with hardening flags applied, any
improvements can be done as a follow-up later.