[Benchmark] detect device in benchmarks (#19327)
and adjust kernel execution time for SubmitKernel
to ensure similar total execution time for BMG
and PVC.

If there is more than one device on the platform,
a warning is printed and the architecture is not set
(which means be will default to some generic benchmark arguments).

This can be extended in future - e.g. we could allow passing a specific
architecture to scripts, similarly as llvm-lit supports `sycl_devices`.