[libclc][Native CPU] Add missing _CLC_DEF. (#19367)
libclc functions are meant to be declared with _CLC_DECL and defined
with _CLC_DEF. A few Native CPU functions were defined without _CLC_DEF,
so add the missing macro.

This is effectively NFC, this applied to trivial functions that were
being defined without always_inline but that would be inlined anyway.