[SYCL][UR] Unify logging and leak checking for L0 v1 and v2 (#19328)
This is needed so that we can enable V2 adapter by default on certain
platforms: https://github.com/intel/llvm/pull/19333

The reason is that we need to load both adapters (legacy and v2) to
check the device version. However, loading v2 adapter causes L0 loader
to emit logs for all API calls (if ZE_DEBUG=1 is set). Since the legacy
adapter used different logic for printing API calls, this would result
in printing the same logs twice. This patch fixes that.