[SYCL][XPTI][NFC] Unify stream type
Data type which identifies an XPTI stream wasn't uniformly the same,
even within XPTI itself different APIs expected different data type as a
stream identifier.

This PR introduces the `xpti::stream_id_t` type alias and uses it
everywhere.

Motivation for this PR is to address warnings produced by
`-Wconversion`. We don't have that flag enabled yet, but it is a
requirement per our SDL processes. It will be enabled in intel/llvm#19306