[SYCL] Drop optimization level when merging properties
Image properties are merged when images are linked. This merging
currently fails if some of the properties conflict with eachother. One
of the properties is the optimization level used when compiling the
binaries, which are used when compiling the images. However, since
linking uses compiled binaries, the optimization level has no effect
on the merged image. As such, this commit drops the property.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>