Mark __sycl_cmath_wrapper_impl.hpp as expected to fail self-contained headers test.
The header is supposed to be moved to the clang headers soon and it can
be included only after cmath.

@npmiller, there are still some issues we might need to address. For
instance, std::is_floating_point, std::enable_if and other type traits
are used w/o including corresponding header from the standard library.
It seems that cmath indirectly includes type_traits, but we probably
should include it explicitly.

There are also use of undeclared identifiers from cmath, but we expect
that this header is included only from cmath wrapper and only after this
snippet:

```c++
// Include real STL <cmath> header - the next one from the include
search
// directories.
// GCC/clang support go through this path.
// MSVC doesn't support "#include_next", so we have to be creative.
// Our header is located in "stl_wrappers/complex" so it won't be picked
by the
// following include. MSVC's installation, on the other hand, has the
layout
// where the following would result in the <complex> we want. This is
obviously
// hacky, but the best we can do...
```