[libclc][test] Really run libspirv tests on all targets (#19395)
Due to the way `add_lit_testsuite` works, the `check-all` target was not
running the libspirv (binding) tests on all targets, instead it was
running it on the last target multiple times in parallel. Besides losing
the coverage for other targets, this also caused sporadic failures due
to tests overwriting each other.
https://github.com/intel/llvm/pull/17902 was an attempt to fix this, and
it works for running all of the `check-libclc-spirv-*` cmake targets in
parallel, but it does not work for the `check-all` target, which is what
we use in the CI.

This patch fixes the issue by making a different `lit` test suite for
each target. See the comment in `libclc/test/CMakeLists.txt` for more
details.