[NFCI][SYCL] Change `device_image_impl::MDevices` to store raw `device_impl *`
https://github.com/intel/llvm/pull/18251 extended `device_impl`s' lifetimes
until shutdown and https://github.com/intel/llvm/pull/18270 started to pass
devices as raw pointers in some of the APIs.

This PR builds on top of that and extends usage of raw
pointers/references/`device_range` as the devices are known to be alive
and extra `std::shared_ptr`'s atomic increments aren't necessary and
could be avoided.

This change mostly touches `device_image_impl` and `program_manager` and
switches most of the APIs to use `devices_range`. A few number of other
modifications are caused by these APIs' changes and are necessary to
keep the code buildable.