[SYCL][NFC] Remove makeResultFilename function from sycl-post-link (#17602)
The functionality of the makeResultFilename function has been moved to
the getOutputPrefix function, and it is invoked at the beginning of the
program.

Two SaveModuleIR functions have been merged into one function.

The output setup in the saveModuleProperties function has been moved out
of the function for simplification.

This simplifying change is mandatory for the later migration of IO
utilities of sycl-post-link to the library.