[SYCL] Bring back RTC support for AMD and Nvidia GPU targets (#19342)
This PR brings back #18918 and #19302, and fixes the issue with shared
library builds.

The problem was that we accessed hidden symbols defined in headers from
the `clang/lib` directory to obtain paths to the vendor-specific device
library files. We now use the `ToolChain::getDeviceLibs` API, and supply
a minimal implementation for the `CudaToolchain`.

---------

Signed-off-by: Julian Oppermann <julian.oppermann@codeplay.com>