[NFCI][SYCL] Change `kernel_bundle_impl::MDevices` to store raw `device_impl *`
https://github.com/intel/llvm/pull/18251 extended `device_impl`s'
lifetimes until shutdown and https://github.com/intel/llvm/pull/18270
started to pass devices as raw pointers in some of the APIs.

This PR builds on top of that and extends usage of raw
pointers/references/`device_range` as the devices are known to be alive
and extra `std::shared_ptr`'s atomic increments aren't necessary and
could be avoided.

Since we change the type of `kernel_bundle_impl::MDevices`, other APIs in
that class don't need to operate in terms of `sycl::device` or
`std::shared_ptr<device_impl>` and we can switch them to use `devices_range`
instead. A small number of other modifications are caused by these APIs' changes
and are necessary to keep the code buildable.