[NFCI][SYCL] Change `device_image_impl::MDevices` to store raw `device_impl *` (#19459)
https://github.com/intel/llvm/pull/18251 extended `device_impl`s'
lifetimes until shutdown and https://github.com/intel/llvm/pull/18270
started to pass devices as raw pointers in some of the APIs.

This PR builds on top of that and extends usage of raw
pointers/references/`device_range` as the devices are known to be alive
and extra `std::shared_ptr`'s atomic increments aren't necessary and
could be avoided.

Since we change the type of `device_image_impl::MDevices`, other APIs in
that class and in `program_manager` don't need to operate in terms of
`sycl::device` or `std::shared_ptr<device_impl>` and we can switch them
to use `devices_range` instead. A small number of other modifications
are caused by these APIs' changes and are necessary to keep the code
buildable.

One extra change is the addition of a minor
`devices_range::to<std::vector<ur_device_handle_t>>()` helper that we
can use now that most of the arguments are `device_range`. Technically,
could go in another PR but then we'd just be modifying the exact same
lines two times, so I decided to fuse it here.