[Driver][SYCL] Update sycl lib linking with -fms-runtime-lib (#19380)
The usage of `-fms-runtime-lib=val` will pull in the default libraries
for MSVC linking as well as the dependent SYCL libraries when using the
Linux based driver with the MSVC target triple. There are special
interactions with the --dependent-lib=msvcrtd library we have that
control the inclusion of sycl.lib to prevent linktime problems.

Extend these behaviors to usage of `-fms-runtime-lib=dll_dbg`,
preventing duplicate instances of the sycl.lib in the --dependent-lib
usage as well as when performing the final link.