[SYCL][UR][L0 v2] move ZeUSMImport definition to adapter.cpp
ZeUSMImport is used by global adapter constructor (on Windows).
It needs to be initialized before the global adapter.

According to Coverity:
The constructor of global object "GlobalAdapter" itself makes use
of global object "ZeUSMImport" defined in another compilation unit.
The order of construction is unspecified, so "GlobalAdapter" might
be created before "ZeUSMImport" is available.