[SYCL][Graph] Disable E2E tests on MTL
Mark SYCL-Graph E2E tests as unsupported on MTL. We have built up several
bug reports on this architecture that need investigated and fixed, see
the code comment.
Rather than tracking lots of individual bugs, disable MTL until we
can give the architecture a thorough investigation and re-enable once
we have fixed the bugs and are confident in its support.

Note that unlike the other architectures in the lit file, these also manifest
on Ubuntu as well as Windows.