[SYCL] Refactor cmath wrapper templates (#19483)
Follow up from: https://github.com/intel/llvm/pull/18706

* Include `<type_traits>`, this was working because `<cmath>` pulls it
in but this is cleaner.
* Use C++14 and C++17 `_v` and `_t` helpers.
* Switch to variadic template for `__sycl_promote`, using C++17 fold
expressions. In theory only needed for 1 to 3 types, but this is an
internal helper anyway so it doesn't seem worth restricting.
* Switch `typedef` to `using`

With this patch the header now require C++17 support to compile. Which
is fine since it's the minimum requirement for SYCL 2020, and this
header is only compiled on the device side which is always handled by
clang, so there shouldn't be any support issues.