[SYCL][Graph] async_malloc use allocation size for zeVirtualMemQueryPageSize (#19402)
In L0 we need to call
[zeVirtualMemQueryPageSize](https://oneapi-src.github.io/level-zero-spec/level-zero/latest/core/api.html#ze__api_8h_1afe411cceb631a3a176463ee05fda3dd7)
with the actual allocation size for the virtual/physical allocations to
align correctly.

Right now we check alignment without passing any size:
https://github.com/intel/llvm/blob/sycl/sycl/source/detail/graph/memory_pool.cpp#L45

This ends up translating to 1 byte in the call to L0:
https://github.com/oneapi-src/unified-runtime/blob/de05f984aa19458a4993d2a2709e3b79d82f1a37/source/adapters/level_zero/virtual_mem.cpp#L32-L37
and for large allocations a wrong alignment is used and L0 reports
ZE_RESULT_ERROR_UNSUPPORTED_SIZE upon zePhysicalMemCreate call (UR fails
with UR_RESULT_ERROR_INVALID_VALUE then).

The UR API should change to accept a size.

This PR exposes this issue in a unittest and fixes it.