[SYCL] Revert m_ExportedSymbolImages key type back to a string (#19478)
It was changed to a string view in preview builds of the library under
the assumption that such strings are always owned by one of the loaded
images.

Bfloat16 device library images are an exception to that rule, where the
dynamic device binary image ends up using a reference to the string from
another binary. If that binary is unloaded, the string view keys will be
invalidated.

Since this particular map lies outside of the kernel submission hotpath
(and it's the only one with a string view key type that gets filled out
for bfloat16 device binary images), this patch just reverts the key type
back to a string.