[Driver][SYCL][New offload] Support external host compiler
Use of -fsycl-host-compiler allows for users to use an external compiler
to perform the host compilation when performing offload builds.  The new
model has the expectation for the device 'object' to be embedded within
the host object via the `clang-offload-packager`.  These fat objects are
consumed by the `clang-linker-wrapper`.

External host compilers do not have the ability to embed the device
binary when creating the host object.  Fortunately, the
`clang-linker-wrapper` allows for consumption of bundled fat objects.

Update the driver when using -fsycl-host-compiler and the new offloading
model to generate bundled files instead of embedded objects.  This
support is for object generation only at this time.  Preprocessing and
assembly file intermediates are not yet supported.