[SYCL][E2E] Avoid illegal failure order in compare_exchange_strong test
The AtomicRef/compare_exchange_strong E2E test allowed `acq_rel` and
`release` being passed as the failure memory order. This is not allowed
by the SYCL AtomicRef spec (see Table 132), and might cause failures on
some backends.

This change explicitly sets the failure order argument in
compare_exchange_strong, instead of relaying on an overload function,
which used the same order for failure and success.