[SYCL][Unit] Fix unit-tests after intel/llvm#19399 (#19543)
intel/llvm#19399 made layout of `scheduler` class dependent on whether
or not XPTI instrumentation is enabled. This caused issues with layout
mismatch when `scheduler.hpp` is included directly from unit-tests,
because macro enabling XPTI were not aligned between the RT and
unit-tests.

The issue was caught for scheduler unit-tests and fixed in
intel/llvm#19399, but apparently we have more unit-tests which need
scheduler. Internal testing discovered issues with buffer unit-tests.

Therefore, re-applied the fix from intel/llvm#19399 globally to all
unit-tests to prevent issues like this appearing in new and other
existing unit-tests.