[SYCL][ESIMD] Run all passes with O0 opt level (#19554)
In https://github.com/intel/llvm/pull/19411, we simply went for running
all passes regardless of the optimization level. However, the change
made it quite difficult to test a lot of the ESIMD functionalities, and
some tests were even useless after the change.
https://github.com/intel/llvm/pull/19411 --and it's follow up
https://github.com/intel/llvm/pull/19453--, is now reverted, and we're
taking a new approach: we still run all passes with `-O0`, but we add a
new `-force-disable-opt` option to `sycl-post-link` so we can still test
for ESIMD functionalities and keep the tests valid.