[SYCL][E2E] Set a proper requirement for subdevice test (#19528)
This test was XFAILed on bmg and dg2 for Linux and Win (partially but
must be completely).
https://github.com/intel/llvm/issues/18576#issuecomment-2949810180
Currently only PVC has support of 4 CCS that is publicly documented
here:

https://github.com/intel/compute-runtime/blob/master/level_zero/doc/experimental_extensions/MULTI_CCS_MODES.md#interaction-with-affinity-mask
Those HW with more than 1 CCS only support 1,2.
This test was added for verification of sycl-ls output
(82ebc2309f84556390af330fa8be012d91535404).
I believe it is enough to test it on PVC only instead of extending it to
other platforms with complex logic of querying and passing capabilities
to the cmdline.

Signed-off-by: Tikhomirova, Kseniya <kseniya.tikhomirova@intel.com>