[SYCL][NFC] Mark external libc functions. (#19368)
<sycl/detail/defines_elementary.hpp> defines a __DPCPP_SYCL_EXTERNAL
macro for functions that are defined elsewhere. This macro is used both
for functions defined by the platform's system libraries and for
functions defined in libclc/libdevice. For Native CPU, these should be
treated differently: functions defined by the platform's system
libraries should be processed according to the platform's ABI, functions
defined by libclc/libdevice should be processed in a
host-target-independent way. Therefore, this PR introduces a
__DPCPP_SYCL_EXTERNAL_LIBC macro so that libc functions can be tagged as
such.

This PR does not yet make the changes to Native CPU to define
__DPCPP_SYCL_EXTERNAL and __DPCPP_SYCL_EXTERNAL_LIBC differently.