[SYCL][NVPTX][AMDGCN] Fix compilation for cmath long double (#19558)
Make sure to not define `long double` wrappers in the cmath wrapper. We
unintentionally defined these for two or three argument functions with
mixed types. These definitions fail to compile if we try to use them,
because there are no SPIR-V builtins for long double. By not defining
them, overload resolution will pick the "host" versions. We can't make
overload resolution fail, because function bodies that are not called on
the device still need to be semantically valid for the device.

This was discovered by the SYCL CTS, in its reference implementation for
the builtins. That code is never called on the device, but it needs to
compile without errors.

---------

Co-authored-by: Steffen Larsen <steffen.larsen@intel.com>