[SYCL][Doc][L0] Add restrctions to interop API
to align with the v2 adapter implementation.

The v2 adapter operates exclusively on in-order command-lists. This
means that it cannot support arbitrary command queue/handle provided
by the user. With the new restrictions, the adapter can choose to
create a new command-list (with the same properties as the handle
passed by the user) and use that instead.

This means that it is no longer possible to synchronize with the
native handle (e.g. by submitting a barrier) directly. The only mechanism
left is native events.