[XPTI][INFRA] Fixes race issues with Emhash under contention (#19600)
- Under severe concurrent access, rehash in Emhash was being called
during insert and this is documented as a problem in Emhash Readme.
Rehash for the Emhash containers will not have to be called now as
enough space for handling streams is reserved.
- xptiCheckTraceEnabled() is now thread-safe
- Emhash containers now have sizes reserved to accommodate the entries
and all second level hash maps that were Emhash in the past have been
replaced with parallel hashmap.

---------

Signed-off-by: Vasanth Tovinkere <vasanth.tovinkere@intel.com>