[NFCI][SYCL] Change `scheduler::enqueueCommandForCG` to accept `event_impl &` (#19599)
Before this PR, its implementation had code paths for `nullptr`
argument, but it's never `nullptr` at any of the call sites. Originally,
this method was added in https://github.com/intel/llvm/pull/7531 and was
called with `nullptr` argument from some fusion code (at least), but it
doesn't seem to be the case anymore (and fusion support has been removed
from the project).

Also, for some reason it was accepting `std::shared_ptr<event_impl>` by
value and not by reference (no idea why), and the name was `NewEvent`
without any comments/documentation. The former is irrelevant after this
change, and for the latter I'm changing the name to a neutral "Event".
If someone has a better understanding, please let me know.