[NewOffload] Fix compile/link options in clang-linker-wrapper
This patch fixes a handling of compile/link options encoded in images.
Also the patch fixes a redirection to backend compiler and offload
wrapper depending on the Triple (AOT or JIT).

This functionality is important for the support of the following clang
command line options in the New Offload Model:
* -ftarget-compile-fast
* -ftarget-export-symbols
* -ftarget-register-alloc-mode