[SYCL][UR][L0 v2] use UR ioq by default (even for ooo sycl queue)
SYCL queue is created as Out-of-order by default. The performance
of OOO queue in L0 v2 adapter is significantly worse than the
performance of the in-order queue and also worse than OOO queue
in the legacy adapter.

Make IOQ the default choice, even if OOO is requested, unless
user explicitly asks for OOO, by setting an env variable.