[SYCL][LIT] Disable tests that fail with libc++ (#19618)
These tests fail when the compiler is built with libc++.
All have Github issues tracking them.

Some of the tests are gtest and not lit so I had to disable those in
CMake. Using defines isn't safe because the problem is the compiler
being compiled with libc++, not necessarily it currently being used to
compile the tests.

---------

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>