[SYCL][Driver][libdevice] Link libspirv in `-fsycl-device-only` mode (#19136)
<!-- start git-machete generated -->

# Based on PR #19135

## Chain of upstream PRs as of 2025-06-25

* PR #19130

  * PR #19131

    * PR #19134

      * PR #19135

        * **PR #19136 (THIS ONE)**

<!-- end git-machete generated -->

Previously, the `-fsycl-device-only` mode disabled linking libspirv.

`-fsycl-device-only` should produce the same LLVM-IR as `-fsycl` would,
but without any host actions.

Libdevice is the likely reason why this was done as is, but it should
not be the default. This patch changes libdevice options to disable
libspirv linking there, and enables linking by default in
`-fsycl-device-only` mode.