[SYCL] Add flag to print CMake args without execution to buildbot.py
Adds a `--print-cmake-flags` option to the configure.py helper script.

When used, the script will print the generated list of CMake
arguments to stdout and exit immediately. This suppresses all
other output and does not invoke the cmake command.

This is useful for packaging or integrating with build systems
that need/prefer to call cmake manually.