[SYCL][UR][Bindless][Doc] Fix copy docs and implementation. (#19093)
This patch fixes the implementation of bindless image copies.
Previously, source and destination pitch values were not being set
correctly.

This patch also updates the wording around the requirements for
`ext_oneapi_copy`.

A missing requirement was added to the specification. Namely that the
`CopyExtent` parameter in the `ext_oneapi_copy` functions that take it,
must not have `0` values in any of the three dimensions, they must be
greater than or equal to `1`.

The requirements for `ext_oneapi_copy` have also been re-written to
prescribe what the functions expect, instead of providing a list of
cases in which the function may fail. This should hopefully make it
clearer and more prescriptive, rather than saying the copy function may
fail if some condition is not met, we now say that the functions require
that certain conditions be met.

The coverage for sub-region copy testing has also been extended to
prevent future regressions.