Fix UB in UntypedPtrTyMap (#3283)
SmallDenseMap requires sentinel values out of the
`SPIRVStorageClassKind` range. This became a build error after
llvm-project commit `38f82534bbe9 ("Reject out-of-bounds enum sentinels
in DenseMap/DenseSet. (#150308)", 2025-07-24)`.

As the `SPIRVStorageClassKind` enum comes from an external header we
cannot change it to an enum class or add an underlying type. This is
actually solved in the `spirv.hpp11` header, but migrating to that
header requires more widespread changes; just switch to a
`std::unordered_map` for now.

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/db0ec312a9cacf9