[SYCL] Use SPIR-V built-in function call for all targets and add `BuiltIn` to the name (#19359)
Before this PR, SPIR-V built-in is represented as global variable for
SPIR/SPIR-V targets and as function call for other targets in
include/sycl/__spirv/spirv_vars.hpp.

According to
https://github.com/llvm/llvm-project/blob/main/llvm/docs/SPIRVUsage.rst,
SPIR-V built-in variable can be mapped to either function call or global
variable. So function call representation should work for SPIR-V target
as well.
The benefit of choosing function call over global variable is that
function call works better for AOT targets that bypass SPIR-V. It is
easier for these targets to implements SPIR-V built-in function via
built-in modules like libclc rather than special handling for lowering
global variable in a custom pass.

Add `BuiltIn` to the name to align with SPIR-V friendly IR and global
variable name.