[CI] Change `pack_release` -> `release_toolchain_artifact` input (#19646)
It's unused in trunk and is only passed in `sycl-rel-nightly.yml` in
`sycl-rel-*` branches (trunk version of it is an empty dummy stub). Not
sure how future "release" branches are expected to get one.

I think making the change results in better alignment between different
artifacts produced the workflow, no other reasons beyond that.