[NativeCPU] Use host data layout. (#19544)
For compilations without an aux target, the data layout string is not
used by us because it only appears in device libs where it will be
overwritten when the device lib gets merged in.

For compilations with an aux target, the data layout string should be
set to that of the aux target, the aux target may be relying on that.